move procedure functions to `main`
I noticed there was a lot of context switching
if you kept flow procedures isolated from main.

That's why I adapted/moved them to main.py.

By the way, the start menu title is a huge string.
It has to be assigned in a strange-looking way.
But that way, it's not indented.